Allow uploading an image over the api

+++ tests/conftest.py
@@ -1,4 +1,5 @@
 # Standard library imports
+import pathlib
 import shutil
 
 # Third party imports
@@ -13,10 +14,16 @@ TEST_MEDIA_ROOT = settings.BASE_DIR / "tmp"
 
 @pytest.fixture(autouse=True)
 def use_temp_directory_as_media_root_for_tests() -> None:
+    """
+    Store media files in a temporary directory during tests.
+    """
     with override_settings(MEDIA_ROOT=TEST_MEDIA_ROOT):
         yield
 
 
-def pytest_sessionfinish(session, exitstatus) -> None:
-    shutil.rmtree(TEST_MEDIA_ROOT)
-    # super(session, exitstatus)
+def pytest_sessionfinish() -> None:
+    """
+    Delete the temporary media directory made during tests.
+    """
+    if pathlib.Path.is_dir(TEST_MEDIA_ROOT):
+        shutil.rmtree(TEST_MEDIA_ROOT)
